requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
tiktoken>=0.7.0
rich>=13.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
